refactor: extract shared DashboardHeader
