+++ apps/frontend/app/dashboard/business/analytics/page.tsx
@@ -11,7 +11,7 @@ export const metadata: Metadata = {
   title: "Analytics",
 };
 import { AnalyticsHeader } from "@/components/dashboard/business/analytics-header";
-import { DashboardHeader } from "@/components/dashboard/business/dashboard-header";
+import { DashboardHeader } from "@/components/dashboard/shared/dashboard-header";
 import { AnalyticsStatCard } from "@/components/dashboard/business/analytics-stat-card";
 import { EngagementSpendChart } from "@/components/dashboard/business/engagement-spend-chart";
 import { PlatformSplitCard } from "@/components/dashboard/business/platform-split-card";
