+++ apps/frontend/app/dashboard/business/payouts/page.tsx
@@ -1,5 +1,6 @@
 import type { Metadata } from "next";
-import { DashboardHeader } from '@/components/dashboard/business/dashboard-header';
+import { DashboardHeader } from '@/components/dashboard/shared/dashboard-header';
+import type { Metadata } from "next";
 import { PayoutsHeader } from '@/components/dashboard/business/payouts-header';
 
 export const metadata: Metadata = {
