+++ apps/frontend/app/dashboard/creator/inventory/page.tsx
@@ -1,5 +1,6 @@
 import type { Metadata } from "next";
-import { DashboardHeader } from "@/components/dashboard/creator/dashboard-header";
+import { DashboardHeader } from "@/components/dashboard/shared/dashboard-header";
+import type { Metadata } from "next";
 import { ConnectedChannels } from "@/components/dashboard/creator/connected-channels";
 
 export const metadata: Metadata = {
