+++ apps/frontend/app/dashboard/creator/page.tsx
@@ -1,5 +1,5 @@
 import { ClipboardCheck } from "lucide-react";
-import { DashboardHeader } from "@/components/dashboard/creator/dashboard-header";
+import { DashboardHeader } from "@/components/dashboard/shared/dashboard-header";
 import { StatCard } from "@/components/dashboard/creator/stat-card";
 import { EarningsCard } from "@/components/dashboard/creator/earnings-card";
 import { TrustScoreCard } from "@/components/dashboard/creator/trust-score-card";
